Integrate with OSG's SceneView stereo
Add a VR mode to integrate with OpenSceneGraph's SceneView stereo mode.
This can be activated using the environment variable
OSGXR_MODE=SCENE_VIEW, which forces the swapchain mode to SINGLE.

Instead of creating slave cameras, existing cameras are hooked into, and
the SceneView objects of the renderer for those cameras. If the View has
slave cameras, any which render to the FRAME_BUFFER are hooked into. A
stereo matrices callback creates the view and projection matrices for
left and right eyes when requested.

Signed-off-by: James Hogan <james@albanarts.com>