CompositionLayerProjection improvements
Tweak CompositionLayerProjection to size the projection view list at
construction based on a new parameter, and to assert if an out of range
viewIndex is provided to addView(). This should minorly reduce copying
and reallocation.

Also warn in getXr() if the projection views provided were incomplete
(i.e. some view is missing).

Signed-off-by: James Hogan <james@albanarts.com>