Add Manager class to allow tighter integration
Add a Manager class to the public API which applications can extend in
order to implement callbacks and more conveniently call into osgXR
functionality. This will allow tighter integration of view setup when
the basic automatic functionality of <osgXR/osgXR> and
<osgXR/OpenXRDisplay> don't cut it.

This is passed into XRState which doesn't yet use it.

Signed-off-by: James Hogan <james@albanarts.com>