Manager: Add API to allow better rendering setup
Add an application Manager callback to allow osgXR to request the
application configure a new view of the world and inform osgXR of any
slave cameras being added or removed which need hooking into.

This will remove the need for unreliable detection of which slave
cameras to hook into, especially when the application may want to
display certain things (such as a GUI) only in the application window.

It also allows an application to continue to manage its own slave
cameras, as long as it keeps osgXR informed of changes to slave cameras
it wants osgXR to hook into.

Signed-off-by: James Hogan <james@albanarts.com>