XRState: Implement new Manager API
Implement the new Manager API for allowing the application to manage
views of the world. When a Manager is in use, we allow the application
to determine which slave cameras osgXR should hook into rather than
guessing.

We can therefore enable the default swap buffers implementation to be
called so that non-VR content can be drawn to the application window.

Signed-off-by: James Hogan <james@albanarts.com>