Add mirror setup API
Add a public API to the Manager class to allow setup of mirror cameras,
according to the MirrorSettings in Settings.

A new public Mirror class encapsulates some of the details, but can be
easily avoided using Manager::setupMirrorCamera() for now.

Current modes supported are NONE, SINGLE (left or right), and LEFT_RIGHT
(side-by-side).

Signed-off-by: James Hogan <james@albanarts.com>