Handle concurrent frames
Handle concurrent frames being accessed, such as when the next frame is
started for update prior to the previous one finishing rendering.

The single mutex locked XRState::_frame is replaced by a new FrameStore
class which has a simple array to store a couple of frames concurrently,
and methods for safely retrieving, creating, or ending a frame.

Frames are distinguished by osg::FrameStamp::getFrameNumber(), with
osg::FrameStamp pointers being passed around XRState and to FrameStore.

Signed-off-by: James Hogan <james@albanarts.com>