Frame stamp XRSwapchain image framebuffers
Tag XRSwapchain image framebuffers with the frame stamp of the last
frame to acquire it. We do this using a basic new helper template
FrameStampedVector to use osg::FrameStamp outside but frame numbers
internally.

This will allow the correct texture to be retrieved for a specific
recent frame, which will be needed for VR mirroring.

Signed-off-by: James Hogan <james@albanarts.com>