GraphicsBindingX11: Hack to get visualid and FBConfig
Use a bit of a hack to get protected member data from GraphicsWindowX11
that don't have public accessors to put in the OpenXR graphics binding.

A derived class extends GraphicsWindowX11 with accessors for _visualInfo
and _fbConfig, and GraphicsBindingX11::GraphicsBindingImpl() casts the
GraphicsWindowX11 pointer to this derived class in order to access the
protected members.

Signed-off-by: James Hogan <james@albanarts.com>