Make OpenXR bringup reversible and restartable
Rewrite XRState's OpenXR initialisation sequence based on a fairly
simple bidirectional set of steps, allowing it to be incrementally torn
down to a certain step, and initialised to a certain step, on successive
update() calls. This allows osgXR to properly respond to OpenXR session
state change events and to initiate an exit.

The steps in question are listed in the XRState::VRState enum.

The public API is extended with these methods:
 - Manager::update() - should be called regularly. This is where
   initialisation, tear down, and callbacks take place.
 - Manager::updateToDestState() - A convenience method to keep updating
   until the destination state has been reached, which is useful to
   complete proper tear down on shutdown.
 - Manager::getEnabled(), Manager::setEnabled() - used for turning VR on
   and off.
 - Manager::syncSettings() - used to detect changes in settings and
   invoke the appropriate levels of restarts to bring the changes about.
 - Manager::isRunning() - returns true if a VR session is running.

And these virtual callback functions
 - Manager::onRunning(), Manager::onStopped() - Allows the app to
   respond to changes in whether VR is running, e.g. by setting up a
   mirror.
 - Manager::onFocus(), Manager::onUnfocus() - Allows the app to respond
   to VR focus changes, e.g. by pausing / unpausing the experience.

Signed-off-by: James Hogan <james@albanarts.com>