Add API to retrieve user-friendly VR state string
Add an API to Manager/XRState to retrieve a relatively user-friendly
string describing the XRState's VR state machine. This isn't easily
translated due to the number of combinations, but the intention is to
avoid exposing XRState internals to Manager users just in order to allow
the app to do the same thing.

Signed-off-by: James Hogan <james@albanarts.com>