Swapchain: Release GL context after xrCreateSwapchain
Old versions of SteamVR such as linux_v1.14 still have value due to bugs
in recent versions [1], however in this version xrCreateSwapchain
unfortunately sets the context without clearing it. Since this may
happen on a different thread to rendering, we should not leave the
context set, or OpenGL operations on the render thread will fail.

This is worked around by clearing the context if it wasn't set before
the call.

[1] https://github.com/ValveSoftware/SteamVR-for-Linux/issues/422

Signed-off-by: James Hogan <james@albanarts.com>