Implement visibility mask support and NodeMask API
Implement visibility mask support in XRState. This uses the new Session
API to get an osg::Geometry for each VR view per depth clearing camera,
and makes an osg::Geode for it, with a custom state set with lighting
disabled, color mask all false, depth range fixed to 0.0f, and a render
bin with id INT_MIN. This is such that it is rendered first, to the
depth buffer only, forcing the depth to 0 in order to mask future
fragments.

The masks are positioned ABSOLUTE_RF camera space, and retransformed on
slave update to be in Z range of the camera's frustum.

Finally a new Manager::setVisibilityMaskNodeMasks() API allows selection
of left and right eye NodeMasks for use in SceneView mode so that
visibility masks for left and right eyes are only rendered for the
appropriate eye.

Signed-off-by: James Hogan <james@albanarts.com>