XRState: Fix session recreation
Recreate session by dropping run state down to VRSTATE_SYSTEM rather
than VRSTATE_SESSION (which is highest).

This exposes a bug when session exit is requested where it is treated as
an external exit that shouldn't retry session creation. Fix
onSessionStateEnd() to avoid disabling VR if an exit was requested
explicitly due to session recreation.

Signed-off-by: James Hogan <james@albanarts.com>