Revert version to 0.3.3
The recent visibility mask changes are source-backwards-compatible with
previous 0.3 patch versions of osgXR, so didn't warrant a minor version
bump to 0.4. As such revert the version back to 0.3.3, but leave the
soversion at 4.

Version 0.4.0 will be treated as equivalent to 0.3.3, and the next
backwards-incompatible change will bump the version to 0.5.0.

Fixes: ff0b464a5a30 ("Bump version to 0.4.0, soversion 4")
Signed-off-by: James Hogan <james@albanarts.com>