Simplify draw pass accounting & fix slave cam mode
Simplify draw pass accounting so that the correct number of draw passes
is expected for slave camera VR mode, and so the calculation is simpler
between slave camera and scene view VR modes.

A swapchains expected number of draw passes is now used directly, and
incremented / decremented twice in scene view mode (since it gets drawn
twice for a single slave camera). This eliminates the need for
_passesPerView.

Signed-off-by: James Hogan <james@albanarts.com>