Wrap XrSpace in an OpenXR::Space class
Add a new internal OpenXR::Space class to manage an XrSpace handle, and
make use if it wherever an XrSpace is used. We only use the local
reference space so far, which is passed around to a few places.

The class has a constructor to create a reference space, and a currently
unused locate() API to allow the space to be located relative to some
base space at the time of a particular frame.

Signed-off-by: James Hogan <james@albanarts.com>