Add internal Action handling classes
Add classes for handling OpenXR action sets, actions, and interaction
profiles.

Action sets can be attached to sessions, where they can be activated,
deactivated, and sync'd. After that action states can be obtained, from
which either the current value of the action or the action space for a
pose action can be obtained. The space can be used to obtain a location
relative to a reference space at the time of a given frame.

Signed-off-by: James Hogan <james@albanarts.com>