Add public APIs for managing OpenXR actions
Add public API classes for managing OpenXR action inputs and haptics:
 - Action, ActionBoolean, ActionFloat, ActionVector2f, ActionPose and
   ActionVibration: Represent different kinds of XrAction.
 - ActionSet: Groups of actions that can be activated and deactivated
   together.
 - InteractionProfile: Allows default action bindings for a certain
   interaction profile to be suggested.
 - Subaction: A subaction path (or top level user path) groups physical
   interactions, allowing single actions that apply to both hands to be
   handled separately.

The API in include/osgXR/ hides most of the implementation details,
including the lifetime of OpenXR objects and sessions. Internals are in
src/, with additional headers to declare internal structures
corresponding to each public header.

XRState ties the public API objects to OpenXR objects when VR is brought
up, and ensures objects are released when VR is brought down. It also
keeps track of public ActionSet & InteractionProfile objects, the
mapping of string paths to Subaction::Private objects (which unlike the
other Private objects are shared between multiple public Objects).
Finally it handles interaction profile changed events in order to inform
the application of changes (via virtual callbacks in the public
Subaction class which can be implemented by derived classes).

Signed-off-by: James Hogan <james@albanarts.com>