XRState: Drop counting refs to InteractionProfiles
Stop keeping references to InteractionProfile objects in XRState, as it
prevents the app from managing their lifetime, including allowing them
to be destroyed and recreated if the underlying data is reloaded.

Instead keeps bare pointers to the InteractionProfile::Private objects
and update Private to notify XRState of destruction. XRState rarely
needs to reference the public object, so primarily handling the Private
object cleans up the Private::get() dereferencing. The Private
object now keeps a pointer to the public object, and the remaining
public object use (in the return value of getCurrentInteractionProfile)
can use the new getPublic() accessor.

Signed-off-by: James Hogan <james@albanarts.com>