XRState: Add action setup as separate init stage
After the session initialisation stage, add an actions init stage.
Action setup can't really be undone without restarting the session, but
its still helpful to treat it as a separate stage which can be retried
until successful, but shouldn't be retried after success (since action
sets can only be attached to a session once).

This slightly changes the semantics so that actions are set up on the
first update() after action sets and interaction profiles have been
configured, rather than only trying to set them up once when the session
is brought up.

Signed-off-by: James Hogan <james@albanarts.com>