XRState: Drop counting refs to ActionSets
Stop keeping references to ActionSet objects in XRState, as it prevents
the app from managing their lifetime, including allowing them to be
destroyed and recreated if the underlying data is reloaded.

Instead keep bare pointers to the ActionSet::Private objects and update
ActionSet::Private to notify XRState of destruction. XRState never
actually needs to reference the public object, so exclusively handling
the private object cleans up the Private::get() dereferencing as well as
simplifying Private construction.

The XRState pointer stored in ActionSet::Private is turned into an
observer_ptr, since XRState could theoretically be destroyed first. The
ActionSet::Private::getState() is unused, so it can be removed also.

Signed-off-by: James Hogan <james@albanarts.com>