CMake: Handle building as a subproject
Handle building of osgXR as a subproject imported into a subdirectory of
another project.

This case is detected in the osgXR CMakeLists.txt, and results in only a
static osgXR library being built. The parent project should link against
osgXR in this case (rather than the namespaced osgXR::osgXR which could
conflict against system versions of osgXR).

The public BUILD_INTERFACE include directories are fixed to work in a
subproject and include generated headers.

Signed-off-by: James Hogan <james@albanarts.com>