Add Swapchain & SubImage APIs
Add a public API to allow the application to define swapchains for use
in upcoming composition layers.

These aren't a lot of use just yet, but they can be attached to cameras
to allow them to be rendered into, and attached to mirrors to allow the
mirror texture to be automatically updated after they've been rendered
into.

Also included is a SubImage API, for describing a subimage of a
swapchain. This is defined purely in the header and is intended to be
passed by reference rather than reference counted. It mostly just wraps
a Swapchain smart pointer with a few extra parameters, including a
currently unused array index which may be coded up in future.

Signed-off-by: James Hogan <james@albanarts.com>