Swapchain: Disable mipmapping in osg::Texture2Ds
We don't bother building multiple mipmap levels for swapchain image
textures, so disable mipmapping by setting the MIN_FILTER to NEAREST on
osg::Texture2D objects returned by
OpenXR::Swapchain::getImageOsgTexture().

This avoids a warning from OpenGL when the texture is used, and is
required for apitrace to be able to replay a trace of OpenGL calls.

Signed-off-by: James Hogan <james@albanarts.com>