Handle releaseGLObjects on window close
When the window is closed, OSG destroys the GLX context almost
immediately, resulting in crashes on later OpenXR swapchain destruction.

Before the context is destroyed, releaseGLObjects() gets called on
various objects including the initial draw callback, so handle this by
triggering a quick destruction of the OpenXR session handle (without
cleanly exiting the session state machine).

We also need to handle the resulting invalid session in a few places,
specifically OpenXR::Session::syncActions() and destruction of
OpenXR::Space and OpenXR::Swapchain.

This does also result in another XRState::update() call, even after
threading has been stopped, so be sure not to re-enable it unless it was
already running.

Signed-off-by: James Hogan <james@albanarts.com>