Bump version to 0.5.0 and extend CHANGELOG
Bump version number to 0.5.0 to indicate a source API breakage (due to
renaming of enumerations which apparently were hitting preprocessor
namespace pollution on Windows), and extend CHANGELOG.md with a
description of changes.

The reason for jumping from 0.3 to 0.5 is that a previous jump to 0.4
was aborted as unnecessary so development continued on 0.3.

Also update the example CMake code in README.md to require 0.5.0.

Signed-off-by: James Hogan <james@albanarts.com>