GraphicsBindingWin32.h: Fix missing IUnknown
The openxr_platform.h makes use of IUnknown when XR_USE_PLATFORM_WIN32
is defined, however it leaves it to the user to ensure it is defined.
This can result in compile errors with Microsoft Visual Studio Community
2019 version 16.11.18, and OpenXR-SDK 1.0.24:
  openxr_platform.h(511,100): error C2061: syntax error: identifier 'IUnknown'

This is fixed by including <unknwn.h> from GraphicsBindingWin32.h, which
is the only place that osgXR includes openxr_platform.h with
XR_USE_PLATFORM_WIN32.

Reported-by: Alan Teeder <ajteeder@v-twin.org.uk>
Tested-by: Alan Teeder <ajteeder@v-twin.org.uk>
Signed-off-by: James Hogan <james@albanarts.com>