Add Hand scenegraph node API
Add a public API scenegraph node class for drawing hands in a chosen
pose (such as from a HandPoseTracked object). Its real simple right now.
Appropriately sized capsules are drawn for each joint in the hand, and
the pose object is passed to the constructor.

FIXME: The HandPose is updated on each update, which could become
redundant if updated elsewhere too.

FIXME: This will likely be excluded from the official API in favour of
letting the APP handle this in its own way with its own shaders.

Signed-off-by: James Hogan <james@albanarts.com>