XRState: Kill un-begun frames at end
If the session is ended abruptly mid-frame, endFrame() may be called
without the frame being begun. In this case we still want to destroy the
frame so that there aren't dangling references to the session, so
explicitly kill the frame.

Also add an assert in downSession() that there should be no frames left,
otherwise the session reference would prevent the session from being
destroyed.

Signed-off-by: James Hogan <james@albanarts.com>