Handle instance loss while session running
An instance loss while a session is running should be treated similarly
to a session loss in terms of bringing down the session safely.

Monado currently reports XR_ERROR_SESSION_LOST from the session
functions even after an XR_ERROR_INSTANCE_LOST error, so we are highly
likely to miss this case at present, however I believe a conformant
runtime would not do this [1], so lets handle that situation too.

[1]: https://gitlab.freedesktop.org/monado/monado/-/issues/319

Signed-off-by: James Hogan <james@albanarts.com>