Clean up extension function pointers
Clean up a few things about how OpenXR extension function pointers are
handled in the Instance and Session classes. The wrapper functions can
all be const including xrGetVisibilityMask(), the pointers themselves
can be safely initialised to nullptr to ensure misuse in absence of the
relevant extension is handled, and they don't need to be mutable since
they aren't obtained on demand. Also clean up the indentation in the
Session wrapper of xrGetVisibilityMask()

Signed-off-by: James Hogan <james@albanarts.com>