View: Expose MVR FB width/height & buf cells
Expose the width and height that full size intermediate frame buffers
should be via View::getMVRWidth() and View::getMVRHeight(). This will
allow apps to create the correct size of intermediate frame buffers to
encompass all subviews that are needed for single-pass multiview render
passes.

Also expose the number of 2D cells to be used for intermediate
fixed-size buffers in order to have data for every subview. This will
allow apps to scale intermediate fixed-size buffers to encompass all
subviews in a similar way.

Signed-off-by: James Hogan <james@albanarts.com>