Add support for layered (array) swapchain images
Add a swapchain mode for layered (2d-array) swapchain images. This
generally does a nicer job of separating the views compared to
side-by-side split, so long as they are similarly sized. However the
primary motivation is to allow support for the GL_OVR_multiview
extension for accelerating multiview rendering, since that extension
requires the frame buffer attachments to be layered.

The osgXR::View class now exposes the following information:
 - The number of layers that MVR frame buffers and fixed-size buffers
   should have, via View::getMVRLayers().
 - The face value to pass to OpenScenGraph when attaching MVR frame
   buffers and fixed-size buffers to a camera for rendering, via
   View::getMVRAttachmentFace(). This allows OSG to be told to attach
   multiple layers to the FrameBuffer object in such a way as to allow
   the layer written to be chosen by some specific mechanism, for
   example the gl_Layer builtin from a geometry shader, or the view ID
   from a vertex shader supporting GL_OVR_multiview.
 - Strings containing GLSL code for each shader stage specifying the
   current layer that should be used when sampling intermediate MVR
   frame buffers or fixed-size buffers, via View::getMVRLayerStr().

The desktop mirror shaders are adapted to pass the layer number to use
when sampling the swapchain images via the Z vertex coordinates, which
allows us to continue to use osg::createTexturedQuadGeometry() which
only accepts 2D texture coordinates.

Signed-off-by: James Hogan <james@albanarts.com>