Add geometry shaders VR mode
Add a VR mode that performs multiview rendering using geometry shaders
provided by the application.

The MultiView class is used to create a shared projection frustum
encompassing all views, which is used for culling.

Indexed viewports are set up using a common function
AppView::setupIndexedViewports() which can be shared with
GL_OVR_multiview support. This allows the viewport to be decided by
index in the geometry shader.

This mode requires that geometry shaders exist in the app for all MVR
content, conditional upon the define OSGXR_GEOM. Views are indexed using
gl_InvocationID in geometry shaders, and gl_ViewportIndex in fragment
shaders. When layered swapchains are enabled the layer is determined by
gl_Layer from geometry and fragment shaders.

Definitions proivided in this mode:
 - OSGXR_VERT_TRANSFORM(POS) (MVR passes): Transforms model/local space
   vertices into shared view space (instead of the usual projection
   space) in vertex shaders, ready for a geometry shader to handle
   per-view transform.
 - OSGXR_GEOM_TRANSFORM(POS) (MVR passes): Transforms a shared view
   space vertex provided by the vertex shader into per-view projection
   space.
 - OSGXR_GEOM_GLOBAL: Sets up the instancing and declares uniforms in
   geometry shader.
 - OSGXR_GEOM_PREPARE_VERTEX: Prepare a vertex for the appropriate view,
   i.e. setting gl_ViewportIndex and gl_Layer appropriately.
 - OSGXR_GEOM_VIEW_MATRIX (MVR passes): Matrix to transform a shared
   view space position vector into per-view space. Similar to
   OSGXR_VERT_VIEW_MATRIX.
 - OSGXR_GEOM_NORMAL_MATRIX (MVR passes): Matrix to transform a shared
   view space normal vector into per-view space. Similar to
   OSGXR_VERT_NORMAL_MATRIX.
 - OSGXR_GEOM_MVR_TEXCOORD (MVR shading passes): Same as
   OSGXR_VERT_MVR_TEXCOORD in other modes, which is undefined (treated
   as no-op) with geometry shaders since the view isn't yet known from
   vertex shaders.
 - OSGXR_GEOM_MVB_TEXCOORD (MVR shading passes): Same as
   OSGXR_VERT_MVB_TEXCOORD in other modes, which is undefined (treated
   as no-op) with geometry shaders since the view isn't yet known from
   vertex shaders.
 - OSGXR_FRAG_GLOBAL (MVR shading passes): As defined for other modes.
 - OSGXR_FRAG_MVR_TEXCOORD (MVR shading passes): As defined for other
   modes.
 - OSGXR_FRAG_MVB_TEXCOORD (MVR shading passes): As defined for other
   modes.

XRFramebuffer is adapted to allow a whole layered texture to be attached
to a framebuffer with the layer specified by the geometry shader by the
gl_Layer builtin, and XRSwapchain is tweaked to allow only a single
framebuffer to be created in this case even with multiple layers.

Geometry shader mode can only be enabled explicitly for now, due to the
required shader additions and changes.

Signed-off-by: James Hogan <james@albanarts.com>