Settings: Add view alignment mask setting
Add app configurable setting for view alignment mask, which controls the
alignment of viewports in shared swapchains. This will allow the app to
keep sampling of intermediate frame buffers within a view when those
buffers are a power of two fraction of the size of the main swapchain
frame buffer.

Signed-off-by: James Hogan <james@albanarts.com>