Settings: Change VR & swapchain mode to prefer/allow ideom
Change the VR mode and swapchain mode selection in osgXR::Settings to
use the prefer/allow ideom used for other settings such as environment
blend modes. This allows the application to be explicit about multiple
modes which are supported, and which modes it would prefer to be chosen.

The existing setVRMode(), getVRMode(), setSwapchainMode() and
getSwapchainMode() are kept for backwards source compatibility, but are
altered to prefer only the chosen mode, and to return the highest
priority mode which is selected in the prefer or allow mask.

Signed-off-by: James Hogan <james@albanarts.com>