XRState: Improve mode choosing
Choose VR and swapchain modes more algorithmically based on preferred
and allowed VR and swapchain mode masks now that the app can specify
them separately.

This will allow safe choice of modes supported by both hardware and the
application (when app changes are required to support certain modes).

The mode choosing algorithm is now moved from upSession() into
chooseMode(), which encapsulates a mode in a ModePriority, essentially a
single number which encodes the VR mode, swapchain mode, and preference
such that lower values are preferred over higher values. App permitted
ModePriority's are put in a std::set with preferences assigned based on
whether the VR and swapchain mode match the preferred masks. These are
then validated in order with validateMode(), with the first validated
mode being chosen.

This separates the validation and fallback logic, such that validation
simply needs to answer the question "can this mode be supported", and
fallback is handled by ordering of ModePriority's.

Signed-off-by: James Hogan <james@albanarts.com>