Instance: Try OpenXR 1.1, falling back to 1.0
When creating an instance, try first with apiVersion set to
XR_API_VERSION_1_1, and if that fails due to
XR_ERROR_API_VERSION_UNSUPPORTED, fall back to XR_API_VERSION_1_0. The
API version is saved and can be read back later.

This will allow the use features added in OpenXR 1.1 where it is
supported.

Signed-off-by: James Hogan <james@albanarts.com>