Add setup-time Condition API
Add an API to allow the application to specify arbitrary conditions
which need to be evaluated at OpenXR setup time. This will be useful for
example for filtering the interaction profile bindings which are valid
depending on enabled extensions or the API version.

Currently the only concrete type is ConditionApi which is for API
features which are added by an extension and/or an API version.

Signed-off-by: James Hogan <james@albanarts.com>