XRState: Fix context handling in downSystem()
XRState::downSystem() tries to make the GL context active for
destruction of FBOs in XRFramebuffer, however the implementation is
flawed in a couple of ways:
 - It uses OpenXR::Session::makeCurrent(), which is explicitly for the
   purposes of working around broken OpenXR runtimes under X11, rather
   than for normal context management.
 - When threading is disabled, the GL context will already be active,
   and releasing it after clearing the _xrViews will result in later
   failures (usually on FBO creation).

Instead of using OpenXR::Session, use the more generic GraphicsContext
APIs via _window, and only do so if threading was enabled when update()
was called. If not we can assume the context is already active and needs
to remain so.

Signed-off-by: James Hogan <james@albanarts.com>