Create ManagedSpace for managing local ref space
Create a new OpenXR::ManagedSpace for managing the local reference
space, in particular handling pending changes, and allowing for local
recentering. A new XrSpace is created whenever the local space offset is
changed, and the appropriate XrSpace is provided based on the XrTime
which is now provided by callers of Session::getLocalSpace().

Signed-off-by: James Hogan <james@albanarts.com>