Implement XR_EXT_user_presence support
Implement support for the XR_EXT_user_presence extension. This provides
explicit events for when user presence changes are detected.

Unfortunately SteamVR doesn't provide a valid session ID in these
events, and so a quirk is required.

The Manager API is extended with a supportsUserPresence() function, and
a onUserPresence() virtual handler to allow the application to determine
if events are expected, and to handle them.

Signed-off-by: James Hogan <james@albanarts.com>